Serialization of Game Objects Beta
-Just finished code for serialization of game data alongside with PhysX
serialization
-Not tested or used yet
-Switched all game object usage to shared_ptr's
-Minor tweaks in many areas regarding usage of the shared_ptrs and
serialization
-Most memory leaks fixed. Last minute changes may have introduced new
ones. Program runs though
TODO:
--Test serialization
--Begin Deserialization

+++ ProBendingx64/ProBendingx64/XMLReader.cpp
@@ -211,7 +211,7 @@ bool XMLReader::GetBoolValue(bool& outVal, bool attributeVal)
 	return false;
 }
 
-bool XMLReader::GetIntValue(int& outVal, bool attributeVal)
+bool XMLReader::GetLongValue(long long& outVal, bool attributeVal)
 {
 	int val;
